rename common to core
