+++ android_tests/lib/android/specs/common/element/window.rb
@@ -1,5 +1,5 @@
-# rake android[common/element/window]
-describe 'common/element/window' do
+# rake android[core/element/window]
+describe 'core/element/window' do
   # rubocop:disable Lint/UnifiedInteger
   t 'window_size' do
     wait do
