+++ ios_tests/lib/common.rb
@@ -1,5 +1,5 @@
 
-# common methods
+# core methods
 def back_click(opts = {})
   opts ||= {}
   search_wait = opts.fetch(:wait, 60 * 1.7)
