+++ ios_tests/lib/ios/specs/common/helper.rb
@@ -1,5 +1,5 @@
-# rake ios[common/helper]
-describe 'common/helper.rb' do
+# rake ios[core/helper]
+describe 'core/helper.rb' do
   def before_first
     screen.must_equal catalog
   end
