+++ ios_tests/lib/run.rb
@@ -1,4 +1,4 @@
-# ios tests have a set of common helper methods
+# ios tests have a set of core helper methods
 require_relative 'common'
 # run file is identical to android
 require_relative '../../android_tests/lib/run'
