+++ lib/appium_lib/core/command.rb
@@ -2,7 +2,7 @@ module Appium
   class Driver
     module Commands
       COMMAND_NO_ARG = {
-        # common
+        # core
         shake:                      [:post, 'session/:session_id/appium/device/shake'.freeze],
         launch_app:                 [:post, 'session/:session_id/appium/app/launch'.freeze],
         close_app:                  [:post, 'session/:session_id/appium/app/close'.freeze],
@@ -26,7 +26,7 @@ module Appium
       }.freeze
 
       COMMAND = {
-        # common
+        # core
         available_contexts:         [:get,  'session/:session_id/contexts'.freeze],
         set_context:                [:post, 'session/:session_id/context'.freeze],
         app_strings:                [:post, 'session/:session_id/appium/app/strings'.freeze],
@@ -52,11 +52,6 @@ module Appium
         end_coverage:               [:post, 'session/:session_id/appium/app/end_test_coverage'.freeze],
         set_network_connection:     [:post, 'session/:session_id/network_connection'.freeze],
         get_performance_data:       [:post, 'session/:session_id/appium/getPerformanceData'.freeze],
-        get_ime_available_engines:  [:get,  'session/:session_id/ime/available_engines'.freeze],
-        get_ime_active_engine:      [:get,  'session/:session_id/ime/active_engine'.freeze],
-        get_ime_activated:          [:get,  'session/:session_id/ime/activated'.freeze],
-        ime_deactivate:             [:post, 'session/:session_id/ime/deactivate'.freeze],
-        ime_activate:               [:post, 'session/:session_id/ime/activate'.freeze],
 
         # iOS
         touch_id:                   [:post, 'session/:session_id/appium/simulator/touch_id'.freeze],
