+++ lib/appium_lib/core/element/window.rb
