+++ lib/appium_lib/core/error.rb
