+++ lib/appium_lib/core/helper.rb
