+++ lib/appium_lib/core/patch.rb
@@ -1,4 +1,4 @@
-require_relative 'version'
+require_relative '../version'
 
 module Appium
   module Common
