+++ lib/appium_lib/core/search_context.rb
