+++ lib/appium_lib/core/wait.rb
