+++ lib/appium_lib/version.rb
